requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
solders>=0.18.0
solana>=0.30.0
python-dotenv>=1.0.0
